Add: Migration & Seeder

+++ mcs_dashboard/database/migrations/2024_04_26_064212_create_data_mcs_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('data_mcs', function (Blueprint $table) {
             $table->id();
-            $table->integer('no_mcs');
+            $table->bigInteger('no_mcs');
             $table->string('nama');
             $table->string('satuan');
             $table->string('kategori');
